requirements.txt
colorama==0.4.6
pillow==12.0.0
pygame==2.6.1
python-magic==0.4.27
python-magic-bin==0.4.14
qrcode==8.2
